Refactor all visualization-navbar components to glimmer
