+++ app/components/visualization/page-setup/navbar/sql-opener.ts
@@ -1,4 +1,4 @@
-import Component from '@ember/component';
+import Component from '@glimmer/component';
 import { inject as service } from "@ember/service";
 import AlertifyHandler from 'explorviz-frontend/utils/alertify-handler';
 import AdditionalData from 'explorviz-frontend/services/additional-data';
@@ -6,9 +6,6 @@ import LandscapeRepository from 'explorviz-frontend/services/repos/landscape-rep
 import { action } from '@ember/object';
 
 export default class SQLOpener extends Component {
-
-  // No Ember generated container
-  tagName = '';
 
   @service('additional-data')
   additionalData!: AdditionalData;
@@ -18,15 +15,15 @@ export default class SQLOpener extends Component {
 
   @action
   showSql() {
-    const latestApplication = this.get('landscapeRepo').get('latestApplication');
+    const latestApplication = this.landscapeRepo.latestApplication;
 
     if(latestApplication !== null) {
-      if (latestApplication.get('databaseQueries').length === 0){
+      if (latestApplication.databaseQueries.length === 0){
         AlertifyHandler.showAlertifyMessage("No SQL statements found!");
         return;
       }
-      this.get('additionalData').addComponent("visualization/page-setup/sidebar/sql-viewer");
-      this.get('additionalData').openAdditionalData();
+      this.additionalData.addComponent("visualization/page-setup/sidebar/sql-viewer");
+      this.additionalData.openAdditionalData();
     }
   }
