+++ app/templates/components/visualization/page-setup/navbar/application-opener.hbs
@@ -1,7 +1,7 @@
 <li class="nav-item mx-2">
-  {{#unless openedActive}}
-  <button class="nav-link btn btn-sm btn-outline-secondary" {{action "openAllComponents"}} type="button">Open All Components</button>
+  {{#unless this.openedActive}}
+  <button class="nav-link btn btn-sm btn-outline-secondary" {{on "click" this.openAllComponents}} type="button">Open All Components</button>
   {{else}}
-  <button class="nav-link btn btn-sm btn-outline-secondary" {{action "closeAllComponents"}} type="button">Close All Components</button>
+  <button class="nav-link btn btn-sm btn-outline-secondary" {{on "click" this.closeAllComponents}} type="button">Close All Components</button>
   {{/unless}}
 </li>
