+++ app/templates/components/visualization/page-setup/navbar/application-search.hbs
@@ -1,21 +1,19 @@
-<li class="nav-item ml-2">
+<li class="nav-item ml-2" {{did-insert this.removePowerselectArrow}}>
 
-    <!-- CSS styling https://ember-power-select.com/cookbook/css-animations -->
+  <!-- CSS styling https://ember-power-select.com/cookbook/css-animations -->
 
-    {{!-- Maybe not the nicest option to use multiple select for single select
-          However, the search inside the trigger seems more intuitive
-     --}}
-    {{#power-select-multiple
-      triggerClass="bootstrap-theme-trigger form-control"
-      search=(perform searchEntity)
-      searchMessage="Insert name or start with '*'"
-      placeholder="Search Entity..." 
-      onchange=(action "onSelect")
-      selected=selectedEntity
-      loadingMessage="loading..."
-      matchTriggerWidth=false
-      as |entity|
-    }}
-      {{entity.name}}
-    {{/power-select-multiple}}
+  {{!-- Maybe not the nicest option to use multiple select for single select
+        However, the search inside the trigger seems more intuitive
+    --}}
+  <PowerSelectMultiple
+    @triggerClass="bootstrap-theme-trigger form-control"
+    @search={{perform this.searchEntity}}
+    @searchMessage="Insert name or start with '*'"
+    @placeholder="Search Entity..."
+    @onchange={{fn this.onSelect}}
+    @loadingMessage="loading..."
+    @matchTriggerWidth={{false}}
+    as |entity|>
+    {{entity.name}}
+  </PowerSelectMultiple>
 </li>
