+++ app/templates/components/visualization/page-setup/navbar/event-opener.hbs
@@ -1,5 +1,5 @@
 <li class="nav-item">
-  <a class="nav-link-with-cursor" title="Show Events" {{action "showEvents"}} >
+  <a class="nav-link-with-cursor" title="Show Events" {{on "click" this.showEvents}} >
     {{svg-jar "bell" class="octicon align-middle"}}
   </a>
 </li>
