+++ app/templates/components/visualization/page-setup/navbar/landscape-downloader.hbs
@@ -1,5 +1,5 @@
 <li class="nav-item">
-    <a class="nav-link-with-cursor" title="Download landscape" {{action "downloadLandscape" landscape}}>
-        {{svg-jar "cloud-download" class="octicon align-middle"}}
-    </a>
+  <a class="nav-link-with-cursor" title="Download landscape" {{on "click" (fn this.downloadLandscape @landscape)}}>
+    {{svg-jar "cloud-download" class="octicon align-middle"}}
+  </a>
 </li>
