+++ app/templates/components/visualization/page-setup/navbar/pause-reload.hbs
@@ -1,11 +1,11 @@
 <li class="nav-item">
 
-  {{#if landscapeListener.pauseVisualizationReload}}
-  <a class="nav-link-with-cursor" title="Resume Visualization" {{action "toggleVisualizationReload"}}>
+  {{#if this.landscapeListener.pauseVisualizationReload}}
+  <a class="nav-link-with-cursor" title="Resume Visualization" {{on "click" this.toggleVisualizationReload}}>
     {{svg-jar "history" class="octicon align-middle navbar-highlight"}}
   </a>
   {{else}}
-  <a class="nav-link-with-cursor" title="Pause Visualization" {{action "toggleVisualizationReload"}}>
+  <a class="nav-link-with-cursor" title="Pause Visualization" {{on "click" this.toggleVisualizationReload}}>
     {{svg-jar "history" class="octicon align-middle"}}
   </a>
   {{/if}}
