+++ app/templates/components/visualization/page-setup/navbar/reset-visualization.hbs
@@ -1,5 +1,5 @@
 <li class="nav-item">
-  <a class="nav-link-with-cursor" title="Reset View" {{action "resetView"}}>
+  <a class="nav-link-with-cursor" title="Reset View" {{on "click" this.resetView}}>
     {{svg-jar "x" class="octicon align-middle"}}
   </a>
 </li>
