+++ app/templates/components/visualization/page-setup/navbar/sql-opener.hbs
@@ -1,5 +1,5 @@
 <li class="nav-item">
-  <a class="nav-link-with-cursor" title="Show Database Queries" {{action "showSql"}} >
+  <a class="nav-link-with-cursor" title="Show Database Queries" {{on "click" this.showSql}} >
     {{svg-jar "database" class="octicon align-middle"}}
   </a>
 </li>
