+++ app/templates/components/visualization/page-setup/navbar/trace-overview.hbs
@@ -1,5 +1,5 @@
 <li class="nav-item">
-  <a class="nav-link-with-cursor" title="Show Traces" {{action "showTraces"}}>
+  <a class="nav-link-with-cursor" title="Show Traces" {{on "click" this.showTraces}}>
     {{svg-jar "list-ordered" class="octicon align-middle"}}
   </a>
 </li>
