+++ package.json
@@ -34,7 +34,7 @@
     "body-parser": "^1.18.3",
     "bootstrap": "^4.2.1",
     "ember-auto-import": "^1.2.19",
-    "ember-bootstrap": "^3.0.0-rc.5",
+    "ember-bootstrap": "^3.0.2",
     "ember-bootstrap-power-select": "^1.0.0",
     "ember-cli": "^3.12.0",
     "ember-cli-dependency-checker": "^3.1.0",
@@ -63,7 +63,7 @@
     "ember-load-initializers": "^2.0.0",
     "ember-modal-dialog": "^3.0.0-beta.3",
     "ember-octicons": "^1.0.0",
-    "ember-power-select": "^2.3.2",
+    "ember-power-select": "^3.0.4",
     "ember-power-select-blockless": "^0.5.0",
     "ember-print-this": "^1.0.0",
     "ember-qunit": "4.4.1",
